T227 Fixes EventChannels to be proper asset-able ScriptableObjects
