--- Assets/Scripts/Events/EventChannelTypes.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 00799906bba50f843b345373b613348b
