+++ Assets/Scripts/Events/EventChannelTypes/GameStateChangedEventChannel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d7bff854e90ee8d4c865c312d408942a
