+++ Assets/Scripts/Events/EventChannelTypes/IntEventChannel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 068779b07ae16814a9f3a17708343f63
