+++ Assets/Scripts/Events/EventChannelTypes/PlayerEventChannel.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 955a0bddebaa4f03a210fb89891f9a31
+timeCreated: 1760998470
