+++ Assets/Scripts/Events/EventChannelTypes/PlayerMovedEventChannel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2e8a107c173876442a290d364cfcadf5
