+++ Assets/Scripts/Events/EventChannelTypes/StringEventChannel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: caf4b917c819ff54189527c91a60092d
